server/requirements.txt
fastapi
uvicorn
pydantic
requests
Pillow
transformers
FlagEmbedding
peft